First rev of new featureTracker application.
The idea: use features computed by one of a few methods (orb, akaze, brisk..)
to characterize a shape. We 'train' the system via rubber-bboxing the
target.  Now we search new frames for keypoint collections that are
sufficiently similar. The hope was that training would result in a more
robust solution.  Currently a single training sample does *not* produce
a sufficiently robust template for use across the range of frames in our
test set.  We might improve by preprocessing the frames to emphasise our
features. Still, the issues I'm currently seeing seem more to do with
perspective than features.